requirements.txt
httplib2==0.7.7
oauth2==1.5.211
python-twitter==0.8.2
simplejson==2.6.2
wsgiref==0.1.2
RPi.GPIO==0.1.0